requirements.txt
pandas
numpy
google-cloud-storage
kfp
google-cloud-aiplatform
transformers
torch
sacrebleu
